refactored xgboost to create functions for other models
